+++ iscc_schema/seed_isrc.py
@@ -24,14 +24,14 @@ class ISRC(BaseModel):
     """
 
     _default_ld: ClassVar[bool] = False
-    context_: Literal["http://purl.org/iscc/context/0.7.0.jsonld"] = Field(
-        "http://purl.org/iscc/context/0.7.0.jsonld",
+    context_: Literal["http://purl.org/iscc/context/0.8.0.jsonld"] = Field(
+        "http://purl.org/iscc/context/0.8.0.jsonld",
         alias="@context",
         description="The JSON-LD Context URI for ISCC metadata.",
     )
     type_: Literal["ISRC"] = Field("ISRC", alias="@type", description="The type of seed metadata.")
-    schema_: Literal["http://purl.org/iscc/schema/isrc-0.7.0.json"] = Field(
-        "http://purl.org/iscc/schema/isrc-0.7.0.json",
+    schema_: Literal["http://purl.org/iscc/schema/isrc-0.8.0.json"] = Field(
+        "http://purl.org/iscc/schema/isrc-0.8.0.json",
         alias="$schema",
         description="The JSON Schema URI for ISRC seed metadata.",
     )